isolating spoti api services
